Move git commands into core
The aim is to have the other layers depend on these abstractions - both
the app / tests that use the commands, and the adapter(s) that implement
handlers for them will have their dependencies pointing at the core.

Once the LocalGitRepo is a little bit cleaner we can move it into a
separate git adapter package.

cc @jbpros - what do you think about the path for importing the
commands? I had to put all the code into the root folder of the core
package to avoid having `src` in the path...