Span new process to start BackgroundGitRepoWorker #99
This improves performance, as something like a `git fetch` will hog
the IO of the main process.

In theory we could start multiple workers like this, and each one could
carry out a fetch in parallel if needed.

It would be better to do this with dedicated docker containers however.