Fix hanging promise in tests
It seemed like the timeout promise, which loses the race in the case of
the succesful ping test, was taking too long and ended up hanging the
whole test suite.

What this does, I think(!), is to ensure that we wait for both promises
to resolve, but ulimately always resolve to whichever promise won the
race.