Add PendingCommitRef#remoteRef property
This broke during the serialize/deserialize for the BackgroundGitRepo,
so I had to implement a `fromJSON` static method. It uses TinyTypes'
Serializable but it's a bit awkward - the properties don't seem to be
there as advertised in the docs, but they do say this is experimental.

I wonder if it would be neater to take named args in the constructor in
this instance.

Also, I wonder if we should be insting on an interface (that has this
property) in the `GitRepo#push` which would have caught this at
compile-time.