Surface error from provider when ValidateRemote fails
When we test for connection with ls-remote and get a failure, we get all
kinds of different errors from the providers.

This commit models all of them as an InvalidRepoUrl and surfaces the
error text from the provider in the message.

In doing this, I discovered a limitation of bullmq, that you can't raise
custom errors in your jobs and expect them to be of that same type when
they get back over the other side of the wire.

I've build a hack in GitError that works around this, by serializing the
custom error's props in a regular Error's `message` property, as this
(and the `stack`) are the only things BullMQ preserves.

I'm not mad keen on this hack, and there will be limitations with it (we
can only use primitives for error props at the moment; it assumes the custom error
constructor takes no args) but it does at least mean we can use the
`InvalidRepoUrl` error to model this failed state right the way through
our stack.

Thoughts?