Try to fix token
The docs for the plugin are inconsistent with GitHub's docs, The GitHub
docs say there should be a token in `github.token` already, and the
plugin's docs say it will use this as the default.

Also, temporarily change this workflow to run on every push, not sure
how that will play out but it might help to test it.

Relates-to: #262