Silence Koa's default stderr error logging
Added some characterisation tests around the createWebApp function to
pin down how the error handling / logging behaviour of the app works.

Cabin is still a bit of a mystery, and how it behaves in concert with
Koa's default behaviour and any other middleware we write is useful to
characterise.

Next: figure out what Cabin already logs about exceptions, and decide
whether to:

- extend what Cabin logs in the http request log to include all error
  info.
- add a second log message about the uncaught exception.

Relates-to: #287