Use CommitName type instead of string
I've also resolved the TODO in handleShowFile, for now anyway, but
having that method take either a CommitName or a RefName. Even if we
don't use the RefName option for now, at least this models what git
could handle.

I don't think we have a use-case for calling this with branch names yet,
so let's not worry about it.