Output logging level as a string
I did a little work here to change around the interface from the tests
to the logging code, using a `captureStdOut` function instead of using
streams on a pino instance. This is simpler, and crucially allows us to
control the options passed to the pino instance, which we need in order
to be able to control the logging level output. It seems you can't
change this for a child logger.

Ref: #599